perf(explore): isolate search keystrokes, right-size All view; split loaders/filters/results
ExploreSearchField owns keystroke state + 350ms debounce (was re-rendering
the full results tree per keypress, twice over via the copy-pasted block in
the All view); All-view blocks fetch 2x block size instead of 4x50 for 5
rendered items (funding keeps the full window for the client-side
Confirmed-by filter); use-explore.ts loader layer split to
use-explore-loaders.ts; quality filters and results area extracted;
explore.tsx 2131 -> ~1480 lines.
