+++ src/components/explore-page/explore.tsx
@@ -1,30 +1,18 @@
 "use client"
 
-import { useCallback, useEffect, useMemo, useRef, useState } from "react"
+import { useCallback, useMemo, useState } from "react"
 import { usePathname, useRouter, useSearchParams } from "next/navigation"
 import {
   ArrowUpDown,
   ChevronDown,
   ChevronRight,
-  Filter as FilterIcon,
   FolderGit2,
   HandCoins,
   LayoutGrid,
   List as ListIcon,
-  TextSearch,
   Users,
 } from "lucide-react"
-import {
-  DEFAULT_HIDDEN_CERT_LABELS,
-  DEFAULT_HIDDEN_ORG_LABELS,
-  HYPERLABEL_DISPLAY_LABELS,
-  HYPERLABEL_DISPLAY_ORDER,
-  HYPERLABEL_TIERS,
-  type HyperlabelTier,
-} from "@/lib/atproto/labels"
 import CertIcon from "@/components/ui/cert-icon"
-import Input from "@/components/ui/input"
-import Checkbox from "@/components/ui/checkbox"
 import {
   Popover as UiPopover,
   PopoverContent,
@@ -34,16 +22,19 @@ import {
 import LoadingSpinner from "@/components/ui/loading-spinner"
 import Tooltip from "@/components/ui/tooltip"
 import SegmentedControl, { ToggleGroup } from "@/components/ui/segmented-control"
-import EmptyState from "@/components/ui/empty-state"
 import SharedLoadMoreSentinel from "@/components/ui/load-more-sentinel"
-import ActivityCard from "@/components/feed/activity-card"
 import CertListRow from "./cert-list-row"
-import ExploreUserCard from "./explore-user-card"
-import ExploreProjectCard from "./explore-project-card"
 import ProjectListRow from "./project-list-row"
 import AccountListRow from "./account-list-row"
 import FundingReceiptRow, { FundingReceiptHeader } from "./funding-receipt-row"
 import FundingConfirmedByPopover from "./funding-confirmed-by-popover"
+import ExploreSearchField from "./explore-search-field"
+import {
+  EMPTY_SELECTION_SENTINEL,
+  QualityFilterPopover,
+  useQualityFilters,
+} from "./quality-filters"
+import { ResultsArea, sortCerts, sortProjects, sortUsers } from "./explore-results"
 import {
   DEFAULT_CONFIRM_ROLES,
   matchesConfirmedBy,
@@ -51,13 +42,16 @@ import {
 } from "@/lib/atproto/funding-provenance"
 import { useFundingConfirmedBy } from "@/hooks/use-funding-confirmed-by"
 import {
+  EMPTY_DID_SET,
   SUB_OPTIONS,
   defaultFilterForView,
   filtersForView,
   parseSubForKind,
   viewFilterToKindFilter,
+  type Degree,
   type ExploreKind,
   type FilterOption,
+  type ListGalleryView,
   type SortOrder,
 } from "./explore-types"
 import { useExploreData } from "@/hooks/use-explore"
@@ -74,63 +68,16 @@ const SORT_LABEL: Record<SortOrder, string> = {
   alphabetical: "Alphabetical",
 }
 
-/** Sentinel slug for the "no label yet" checkbox that sits at the
- *  end of every labeler popover. Backed by an `includeLabels` /
- *  `excludeLabels` swap at the loader (see comment on
- *  `excludeCertLabels` / `includeCertLabels` below). */
-const UNLABELED_SLUG = "unlabeled" as const
-type UnlabeledSlug = typeof UNLABELED_SLUG
-
-const UNLABELED_LABEL = "Not labeled yet"
-
-/**
- * Orglabeler tier slugs used in the URL `?orgQuality=` param. These are
- * also the exact kebab-case label values the indexer stores (per issue
- * #145), so a slug is sent straight to the indexer as the org `labels` /
- * `excludeLabels` value — no slug↔value mapping needed. */
-const ORG_TIER_SLUGS = ["high-quality", "standard", "likely-test"] as const
-type OrgTierSlug = (typeof ORG_TIER_SLUGS)[number]
-
-const ORG_TIER_DISPLAY_LABEL: Record<OrgTierSlug, string> = {
-  "high-quality": "High quality",
-  standard: "Standard",
-  "likely-test": "Likely test",
-}
-
-/** Default org-quality set when `?orgQuality=` is missing —
- *  everything except the labels listed in DEFAULT_HIDDEN_ORG_LABELS
- *  (today only "likely-test"). Matches the home feed's policy. */
-const DEFAULT_ORG_TIER_SLUGS: readonly OrgTierSlug[] = ORG_TIER_SLUGS.filter(
-  (slug) => !DEFAULT_HIDDEN_ORG_LABELS.includes(slug),
-)
-
 function parseSort(v: string | null): SortOrder {
   if (v === "newest" || v === "oldest" || v === "alphabetical") return v
   return "newest"
 }
 
-type ListGalleryView = "list" | "gallery"
 function parseView(v: string | null): ListGalleryView {
   return v === "gallery" ? "gallery" : "list"
 }
 
-type Degree = 1 | 2 | 3
-
 const ALL_DEGREES: readonly Degree[] = [1, 2, 3] as const
-
-/**
- * Sentinel for an explicitly-empty selection. Without this, a writer
- * that puts `""` into the URL would be normalised away by setUrl
- * (which deletes empty values), and the next read would resolve to
- * the default set — making "deselect all" indistinguishable from "no
- * preference" for the user. We pick `-` because it never collides
- * with a legitimate value across degrees / quality / orgQuality.
- */
-const EMPTY_SELECTION_SENTINEL = "-"
-
-/** Shared empty set for the funding "Confirmed by" third-party axis when
- *  none are selected (avoids re-allocating on every receipt filter pass). */
-const EMPTY_DID_SET: ReadonlySet<string> = new Set<string>()
 
 /** Default funding "Confirmed by" selection — all role buckets, no third
  *  parties — i.e. show only receipts confirmed by the sender, recipient, or
@@ -209,284 +156,6 @@ function isEndorsementFilter(kind: ExploreKind, filter: string): boolean {
  * via the on-page dropdown, which drives `?show=`. {@link ExploreAll}
  * branches that into the three-block layout or a single-category pane.
  */
-/**
- * URL-backed quality-filter state shared by the single-kind view
- * (`ExploreMain`) and the combined All view (`ExploreAllBlocks`). Owns
- * the `?quality=` (cert / Activity Labeler tiers) and `?orgQuality=`
- * (Orglabeler tiers) params, derives the include/exclude label arrays
- * the loaders pass to the indexer, and exposes the toggle/reset
- * handlers + "is default" flags the popover renders against.
- */
-function useQualityFilters() {
-  const pathname = usePathname()
-  const searchParams = useSearchParams()
-  const router = useRouter()
-
-  const setUrl = useCallback(
-    (patch: Record<string, string | null>) => {
-      const params = new URLSearchParams(searchParams?.toString() ?? "")
-      for (const [k, v] of Object.entries(patch)) {
-        if (v === null || v === "") params.delete(k)
-        else params.set(k, v)
-      }
-      const qs = params.toString()
-      router.replace(qs ? `${pathname}?${qs}` : pathname, { scroll: false })
-    },
-    [pathname, searchParams, router],
-  )
-
-  // Cert quality (Activity Labeler tiers) — INCLUDED set, with the
-  // synthetic `unlabeled` sentinel. Missing param = home-feed default
-  // (every non-hidden tier + unlabeled).
-  const qualityParam = searchParams?.get("quality")
-  const qualityIncluded = useMemo<Set<HyperlabelTier | UnlabeledSlug>>(() => {
-    if (qualityParam == null) {
-      return new Set<HyperlabelTier | UnlabeledSlug>([
-        ...HYPERLABEL_TIERS.filter((t) => !DEFAULT_HIDDEN_CERT_LABELS.includes(t)),
-        UNLABELED_SLUG,
-      ])
-    }
-    if (qualityParam === EMPTY_SELECTION_SENTINEL) {
-      return new Set<HyperlabelTier | UnlabeledSlug>()
-    }
-    const valid = new Set<string>([...HYPERLABEL_TIERS, UNLABELED_SLUG])
-    return new Set(
-      qualityParam
-        .split(",")
-        .filter((v): v is HyperlabelTier | UnlabeledSlug => valid.has(v)),
-    )
-  }, [qualityParam])
-  const certIncludeUnlabeled = qualityIncluded.has(UNLABELED_SLUG)
-  const excludeCertLabels = useMemo<readonly string[] | undefined>(
-    () =>
-      certIncludeUnlabeled
-        ? HYPERLABEL_TIERS.filter((t) => !qualityIncluded.has(t))
-        : undefined,
-    [qualityIncluded, certIncludeUnlabeled],
-  )
-  const includeCertLabels = useMemo<readonly string[] | undefined>(
-    () =>
-      certIncludeUnlabeled
-        ? undefined
-        : HYPERLABEL_TIERS.filter((t) => qualityIncluded.has(t)),
-    [qualityIncluded, certIncludeUnlabeled],
-  )
-  const qualityIsDefault = useMemo(() => {
-    const expectedSize =
-      HYPERLABEL_TIERS.length - DEFAULT_HIDDEN_CERT_LABELS.length + 1
-    if (qualityIncluded.size !== expectedSize) return false
-    if (!qualityIncluded.has(UNLABELED_SLUG)) return false
-    for (const t of HYPERLABEL_TIERS) {
-      const shouldBeIncluded = !DEFAULT_HIDDEN_CERT_LABELS.includes(t)
-      if (qualityIncluded.has(t) !== shouldBeIncluded) return false
-    }
-    return true
-  }, [qualityIncluded])
-
-  // Org quality (Orglabeler tiers) — same pattern.
-  const orgQualityParam = searchParams?.get("orgQuality")
-  const orgQualityIncluded = useMemo<Set<OrgTierSlug | UnlabeledSlug>>(() => {
-    if (orgQualityParam == null) {
-      return new Set<OrgTierSlug | UnlabeledSlug>([
-        ...DEFAULT_ORG_TIER_SLUGS,
-        UNLABELED_SLUG,
-      ])
-    }
-    if (orgQualityParam === EMPTY_SELECTION_SENTINEL) {
-      return new Set<OrgTierSlug | UnlabeledSlug>()
-    }
-    const valid = new Set<string>([...ORG_TIER_SLUGS, UNLABELED_SLUG])
-    return new Set(
-      orgQualityParam
-        .split(",")
-        .filter((v): v is OrgTierSlug | UnlabeledSlug => valid.has(v)),
-    )
-  }, [orgQualityParam])
-  const orgIncludeUnlabeled = orgQualityIncluded.has(UNLABELED_SLUG)
-  const excludeOrgLabels = useMemo<readonly OrgTierSlug[] | undefined>(
-    () =>
-      orgIncludeUnlabeled
-        ? ORG_TIER_SLUGS.filter((slug) => !orgQualityIncluded.has(slug))
-        : undefined,
-    [orgQualityIncluded, orgIncludeUnlabeled],
-  )
-  const includeOrgLabels = useMemo<readonly OrgTierSlug[] | undefined>(
-    () =>
-      orgIncludeUnlabeled
-        ? undefined
-        : ORG_TIER_SLUGS.filter((slug) => orgQualityIncluded.has(slug)),
-    [orgQualityIncluded, orgIncludeUnlabeled],
-  )
-  const orgQualityIsDefault = useMemo(() => {
-    if (orgQualityIncluded.size !== DEFAULT_ORG_TIER_SLUGS.length + 1) return false
-    if (!orgQualityIncluded.has(UNLABELED_SLUG)) return false
-    for (const slug of ORG_TIER_SLUGS) {
-      const shouldBeIncluded = DEFAULT_ORG_TIER_SLUGS.includes(slug)
-      if (orgQualityIncluded.has(slug) !== shouldBeIncluded) return false
-    }
-    return true
-  }, [orgQualityIncluded])
-
-  const onResetQuality = useCallback(() => {
-    setUrl({ quality: null, orgQuality: null })
-  }, [setUrl])
-  const onQualityToggle = useCallback(
-    (slug: HyperlabelTier | UnlabeledSlug) => {
-      const next = new Set(qualityIncluded)
-      if (next.has(slug)) next.delete(slug)
-      else next.add(slug)
-      const defaultSlugs = new Set<HyperlabelTier | UnlabeledSlug>([
-        ...HYPERLABEL_TIERS.filter((t) => !DEFAULT_HIDDEN_CERT_LABELS.includes(t)),
-        UNLABELED_SLUG,
-      ])
-      const isDefault =
-        next.size === defaultSlugs.size &&
-        Array.from(defaultSlugs).every((s) => next.has(s))
-      const ordered: (HyperlabelTier | UnlabeledSlug)[] = [
-        ...HYPERLABEL_TIERS.filter((t) => next.has(t)),
-        ...(next.has(UNLABELED_SLUG) ? [UNLABELED_SLUG] : []),
-      ]
-      const value = isDefault
-        ? null
-        : ordered.length === 0
-          ? EMPTY_SELECTION_SENTINEL
-          : ordered.join(",")
-      setUrl({ quality: value })
-    },
-    [qualityIncluded, setUrl],
-  )
-  const onOrgQualityToggle = useCallback(
-    (slug: OrgTierSlug | UnlabeledSlug) => {
-      const next = new Set(orgQualityIncluded)
-      if (next.has(slug)) next.delete(slug)
-      else next.add(slug)
-      const defaultSlugs = new Set<OrgTierSlug | UnlabeledSlug>([
-        ...DEFAULT_ORG_TIER_SLUGS,
-        UNLABELED_SLUG,
-      ])
-      const isDefault =
-        next.size === defaultSlugs.size &&
-        Array.from(defaultSlugs).every((s) => next.has(s))
-      const ordered: (OrgTierSlug | UnlabeledSlug)[] = [
-        ...ORG_TIER_SLUGS.filter((s) => next.has(s)),
-        ...(next.has(UNLABELED_SLUG) ? [UNLABELED_SLUG] : []),
-      ]
-      const value = isDefault
-        ? null
-        : ordered.length === 0
-          ? EMPTY_SELECTION_SENTINEL
-          : ordered.join(",")
-      setUrl({ orgQuality: value })
-    },
-    [orgQualityIncluded, setUrl],
-  )
-
-  return {
-    qualityIncluded,
-    orgQualityIncluded,
-    excludeCertLabels,
-    includeCertLabels,
-    excludeOrgLabels,
-    includeOrgLabels,
-    qualityIsDefault,
-    orgQualityIsDefault,
-    onQualityToggle,
-    onOrgQualityToggle,
-    onResetQuality,
-  }
-}
-
-type QualityFilters = ReturnType<typeof useQualityFilters>
-
-/**
- * The quality-filter popover (trigger + content). `showCertSection`
- * adds the Activity-quality (cert) section above Account quality — true
- * on the certs single-kind view and on the All view (which includes
- * activities); false on accounts/projects single-kind views, where only
- * the author-org tier applies.
- */
-function QualityFilterPopover({
-  q,
-  showCertSection,
-  open,
-  onOpenChange,
-}: {
-  q: QualityFilters
-  showCertSection: boolean
-  open: boolean
-  onOpenChange: (v: boolean) => void
-}) {
-  const filtered =
-    (showCertSection && !q.qualityIsDefault) || !q.orgQualityIsDefault
-  return (
-    <UiPopover open={open} onOpenChange={onOpenChange}>
-      <Tooltip label="Filter by quality">
-        <PopoverTrigger>
-          <button
-            type="button"
-            className={`explore__chrome-btn explore__chrome-btn--icon${
-              filtered ? " explore__chrome-btn--active" : ""
-            }`}
-            aria-label={`Filter by quality${filtered ? " (filtered)" : ""}`}
-          >
-            <FilterIcon size={13} strokeWidth={1.75} aria-hidden />
-          </button>
-        </PopoverTrigger>
-      </Tooltip>
-      <PopoverContent align="end">
-        {showCertSection ? (
-          <>
-            <p className="popover__section-heading">Activity quality</p>
-            {HYPERLABEL_DISPLAY_ORDER.map((tier) => (
-              <div key={tier} className="popover__item popover__item--check">
-                <Checkbox
-                  label={HYPERLABEL_DISPLAY_LABELS[tier]}
-                  checked={q.qualityIncluded.has(tier)}
-                  onChange={() => q.onQualityToggle(tier)}
-                />
-              </div>
-            ))}
-            <div className="popover__item popover__item--check">
-              <Checkbox
-                label={UNLABELED_LABEL}
-                checked={q.qualityIncluded.has(UNLABELED_SLUG)}
-                onChange={() => q.onQualityToggle(UNLABELED_SLUG)}
-              />
-            </div>
-            <hr className="popover__divider" aria-hidden="true" />
-          </>
-        ) : null}
-        <p className="popover__section-heading">Account quality</p>
-        {ORG_TIER_SLUGS.map((slug) => (
-          <div key={slug} className="popover__item popover__item--check">
-            <Checkbox
-              label={ORG_TIER_DISPLAY_LABEL[slug]}
-              checked={q.orgQualityIncluded.has(slug)}
-              onChange={() => q.onOrgQualityToggle(slug)}
-            />
-          </div>
-        ))}
-        <div className="popover__item popover__item--check">
-          <Checkbox
-            label={UNLABELED_LABEL}
-            checked={q.orgQualityIncluded.has(UNLABELED_SLUG)}
-            onChange={() => q.onOrgQualityToggle(UNLABELED_SLUG)}
-          />
-        </div>
-        <hr className="popover__divider" aria-hidden="true" />
-        <button
-          type="button"
-          className="popover__reset-btn"
-          onClick={q.onResetQuality}
-          disabled={(!showCertSection || q.qualityIsDefault) && q.orgQualityIsDefault}
-        >
-          Reset to default
-        </button>
-      </PopoverContent>
-    </UiPopover>
-  )
-}
-
 export default function Explore() {
   // Register the page title in the top bar's title slot — mirrors the
   // convention every other top-level page uses (Apps, Settings…).
@@ -586,6 +255,14 @@ function ExploreMain({
     },
     [pathname, searchParams, router],
   )
+
+  // Declared before the callbacks below that close over the setters —
+  // the react-hooks lint rejects accessing a state variable above its
+  // declaration.
+  const [sortOpen, setSortOpen] = useState(false)
+  const [qualityOpen, setQualityOpen] = useState(false)
+  const [confirmedByOpen, setConfirmedByOpen] = useState(false)
+  const [subPrefixOpen, setSubPrefixOpen] = useState(false)
 
   // Read the target's `data-*` attribute instead of capturing the
   // iteration variable in a closure. The SWC minifier (Next 16's
@@ -690,37 +367,13 @@ function ExploreMain({
     [setUrl],
   )
 
-  // Local search debounce: keep typing snappy, hit indexer once typing stops.
-  const [localQuery, setLocalQuery] = useState(search)
-  // Remember the value we last wrote to the URL so the URL→local
-  // sync below can tell our own debounce writes apart from external
-  // URL changes (back/forward, filter switch that clears `q`). Without
-  // this, the sync effect fires every time we write — and if the user
-  // typed an extra keystroke between scheduling the write and the URL
-  // commit, that keystroke gets stomped (it shows on screen briefly,
-  // then the URL→local sync overwrites localQuery with the older URL
-  // value). Symptom: "not all keystrokes are recognised when results
-  // come in."
-  const lastWroteToUrlRef = useRef<string | null>(null)
-  useEffect(() => {
-    if (search === lastWroteToUrlRef.current) return
-    setLocalQuery(search)
-  }, [search])
-  useEffect(() => {
-    const t = setTimeout(() => {
-      if (localQuery !== search) {
-        lastWroteToUrlRef.current = localQuery
-        setUrl({ q: localQuery || null })
-      }
-    }, 350)
-    return () => clearTimeout(t)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [localQuery])
-
-  const [sortOpen, setSortOpen] = useState(false)
-  const [qualityOpen, setQualityOpen] = useState(false)
-  const [confirmedByOpen, setConfirmedByOpen] = useState(false)
-  const [subPrefixOpen, setSubPrefixOpen] = useState(false)
+  // Debounced search — ExploreSearchField owns the keystroke state +
+  // URL sync so a keystroke re-renders only the input, not this whole
+  // chrome + results tree.
+  const onSearchCommit = useCallback(
+    (q: string | null) => setUrl({ q }),
+    [setUrl],
+  )
 
   return (
     <div className="explore__main">
@@ -744,19 +397,11 @@ function ExploreMain({
               />
             ) : null}
 
-            <div className="explore__search-field">
-              <Input
-                type="search"
-                size="sm"
-                leadingIcon={
-                  <TextSearch size={14} strokeWidth={1.75} aria-hidden />
-                }
-                placeholder={searchPlaceholder(kind)}
-                value={localQuery}
-                onChange={(e) => setLocalQuery(e.target.value)}
-                aria-label={searchPlaceholder(kind)}
-              />
-            </div>
+            <ExploreSearchField
+              search={search}
+              placeholder={searchPlaceholder(kind)}
+              onCommit={onSearchCommit}
+            />
 
             {mobilePillsShow !== undefined && mobilePillsSetShow ? (
               <AllCategoryPills
@@ -1007,8 +652,9 @@ function ExploreFilterSidebar({
 
 // ----------------------------- All view -------------------------------------
 
-/** How many results each block on the All view shows. The loader still
- *  fetches a full page per kind; we just render the head of each list. */
+/** How many results each block on the All view shows. The loaders fetch
+ *  a 2x window per kind (`pageSize` below) — enough to fill a block after
+ *  client-side trimming without pulling a full 50-row page for 5 rows. */
 const ALL_VIEW_BLOCK_SIZE = 5
 
 function isValidViewFilter(filter: string): boolean {
@@ -1142,25 +788,12 @@ function ExploreAllBlocks() {
   const quality = useQualityFilters()
   const [qualityOpen, setQualityOpen] = useState(false)
 
-  // Search debounce — same pattern as the single-kind view. Keeps
-  // typing snappy, hits the indexer once typing stops, and reconciles
-  // external URL changes (filter switch / back-forward) into the input.
-  const [localQuery, setLocalQuery] = useState(search)
-  const lastWroteToUrlRef = useRef<string | null>(null)
-  useEffect(() => {
-    if (search === lastWroteToUrlRef.current) return
-    setLocalQuery(search)
-  }, [search])
-  useEffect(() => {
-    const t = setTimeout(() => {
-      if (localQuery !== search) {
-        lastWroteToUrlRef.current = localQuery
-        setUrl({ q: localQuery || null })
-      }
-    }, 350)
-    return () => clearTimeout(t)
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [localQuery])
+  // Debounced search — ExploreSearchField owns the keystroke state so
+  // typing re-renders the input alone, not the four section blocks.
+  const onSearchCommit = useCallback(
+    (q: string | null) => setUrl({ q }),
+    [setUrl],
+  )
 
   // Three independent loaders — one per kind. Each maps the unified
   // All-view filter to that kind's concrete filter key. `sub` is pinned
@@ -1170,6 +803,9 @@ function ExploreAllBlocks() {
     filter: viewFilterToKindFilter(filter, "activities"),
     sub: "all",
     search,
+    // 2x the block cap — a short server page still fills the block
+    // after any client-side trimming, without fetching 50 rows for 5.
+    pageSize: ALL_VIEW_BLOCK_SIZE * 2,
     // Activities carry both their own cert tier and their author org's
     // tier, so both quality axes apply.
     excludeCertLabels: quality.excludeCertLabels,
@@ -1182,6 +818,7 @@ function ExploreAllBlocks() {
     filter: viewFilterToKindFilter(filter, "projects"),
     sub: "all",
     search,
+    pageSize: ALL_VIEW_BLOCK_SIZE * 2,
     // Projects + accounts filter by the author org's tier only.
     excludeOrgLabels: quality.excludeOrgLabels,
     includeOrgLabels: quality.includeOrgLabels,
@@ -1191,12 +828,16 @@ function ExploreAllBlocks() {
     filter: viewFilterToKindFilter(filter, "accounts"),
     sub: "all",
     search,
+    pageSize: ALL_VIEW_BLOCK_SIZE * 2,
     excludeOrgLabels: quality.excludeOrgLabels,
     includeOrgLabels: quality.includeOrgLabels,
   })
   // Funding has no social-graph / featured filters and no quality axis;
   // the loader lists receipts gated to an AT Protocol account on either
   // side, so it always runs with the plain "all" filter.
+  // Keeps the default PAGE_SIZE window (no pageSize cap): the fixed
+  // "Confirmed by" filter below is client-side and can drop most
+  // receipts, so a trimmed page could leave the block short.
   const funding = useExploreData({
     kind: "funding",
     filter: "all",
@@ -1252,17 +893,11 @@ function ExploreAllBlocks() {
               />
             </span>
 
-            <div className="explore__search-field">
-              <Input
-                type="search"
-                size="sm"
-                leadingIcon={<TextSearch size={14} strokeWidth={1.75} aria-hidden />}
-                placeholder="Search all of Explore…"
-                value={localQuery}
-                onChange={(e) => setLocalQuery(e.target.value)}
-                aria-label="Search all of Explore…"
-              />
-            </div>
+            <ExploreSearchField
+              search={search}
+              placeholder="Search all of Explore…"
+              onCommit={onSearchCommit}
+            />
 
             <AllCategoryPills show="all" setShow={setShow} />
 
@@ -1832,300 +1467,4 @@ function searchPlaceholder(kind: ExploreKind): string {
   if (kind === "projects") return "Search projects…"
   if (kind === "funding") return "Funding receipts"
   return "Search activities…"
-}
-
-/** Render whatever the data hook returned, applying client-side sort
- *  and routing through the right card. */
-function ResultsArea({
-  kind,
-  data,
-  sort,
-  view,
-  degrees,
-  confirmRoles,
-  confirmThirdParties,
-}: {
-  kind: ExploreKind
-  data: ReturnType<typeof useExploreData>
-  sort: SortOrder
-  view: ListGalleryView
-  /** Non-null only when the active filter is endorsement-based.
-   *  When present, rows whose author's degree isn't in the set are
-   *  filtered out — the loader fetched the full closure up to
-   *  `max(degrees)`, this trims the subset the user actually wants
-   *  to see. */
-  degrees: Set<Degree> | null
-  /** Funding only — the selected "Confirmed by" role buckets + third-party
-   *  attestor DIDs. Receipts are filtered to the union; with both empty,
-   *  nothing shows. */
-  confirmRoles?: ReadonlySet<ConfirmRole>
-  confirmThirdParties?: ReadonlySet<string>
-}) {
-  const closure = data.endorsementClosure
-  const degreeMatches = useCallback(
-    (did: string | null | undefined): boolean => {
-      if (!degrees || !closure) return true
-      if (!did) return false
-      const meta = closure.closureByDid.get(did)
-      if (!meta) return false
-      return degrees.has(meta.degree)
-    },
-    [degrees, closure],
-  )
-
-  // Funding "Confirmed by" filter — memoized so an unrelated re-render (a
-  // keystroke in search, a view toggle) doesn't re-run the O(n) attestation
-  // filter over the whole receipt list. Recomputes only when the loaded
-  // receipts or either selection changes.
-  // Merge optimistic confirmations + collapse matchingReceipt pairs (issue
-  // #186) before applying the "Confirmed by" filter.
-  const mergedFundingReceipts = useMergedFunding(data.fundingReceipts)
-  const filteredFundingReceipts = useMemo(
-    () =>
-      confirmRoles
-        ? mergedFundingReceipts.filter((r) =>
-            matchesConfirmedBy(
-              r.attestations,
-              confirmRoles,
-              confirmThirdParties ?? EMPTY_DID_SET,
-            ),
-          )
-        : mergedFundingReceipts,
-    [mergedFundingReceipts, confirmRoles, confirmThirdParties],
-  )
-
-  // Degree-filtered + sorted lists, memoized so a keystroke in the search
-  // box (local state on the parent) doesn't re-allocate and re-sort the
-  // whole list each render. The underlying arrays are stable references
-  // between keystrokes (they live in useExploreData's state), so these
-  // recompute only when the loaded data, the active degree set, or the
-  // sort order actually changes.
-  const sortedUsers = useMemo(() => {
-    const actors = degrees
-      ? data.users.filter((a) => degreeMatches(a.did))
-      : data.users
-    return sortUsers(actors, sort)
-  }, [data.users, degrees, degreeMatches, sort])
-  const sortedProjects = useMemo(() => {
-    const list = degrees
-      ? data.projects.filter((p) => degreeMatches(projectAuthorDid(p)))
-      : data.projects
-    return sortProjects(list, sort)
-  }, [data.projects, degrees, degreeMatches, sort])
-  const sortedCerts = useMemo(() => {
-    const list = degrees
-      ? data.certs.filter((c) => degreeMatches(data.certDids.get(c.uri) ?? null))
-      : data.certs
-    return sortCerts(list, sort)
-  }, [data.certs, data.certDids, degrees, degreeMatches, sort])
-
-  if (
-    data.isLoading &&
-    data.users.length === 0 &&
-    data.projects.length === 0 &&
-    data.certs.length === 0 &&
-    data.fundingReceipts.length === 0
-  ) {
-    return (
-      <div className="explore__loading">
-        <LoadingSpinner size="md" />
-      </div>
-    )
-  }
-
-  if (kind === "funding") {
-    const receipts = filteredFundingReceipts
-    if (receipts.length === 0) return <EmptyResults kind={kind} />
-    return (
-      <ul className="explore__list explore__list--funding">
-        <li>
-          <FundingReceiptHeader />
-        </li>
-        {receipts.map((r) => (
-          <li key={r.uri}>
-            <FundingReceiptRow receipt={r} showTextParties />
-          </li>
-        ))}
-      </ul>
-    )
-  }
-
-  if (kind === "accounts") {
-    const actors = sortedUsers
-    if (actors.length === 0) return <EmptyResults kind={kind} />
-    if (view === "list") {
-      return (
-        <ul className="explore__list explore__list--accounts">
-          {actors.map((a) => (
-            <li key={a.did}>
-              <AccountListRow
-                actor={a}
-                endorsementMeta={closure?.closureByDid.get(a.did)}
-              />
-            </li>
-          ))}
-        </ul>
-      )
-    }
-    return (
-      <ul className="explore__grid explore__grid--users">
-        {actors.map((a) => (
-          <li key={a.did}>
-            <ExploreUserCard actor={a} />
-          </li>
-        ))}
-      </ul>
-    )
-  }
-
-  if (kind === "projects") {
-    const projects = sortedProjects
-    if (projects.length === 0) return <EmptyResults kind={kind} />
-    if (view === "list") {
-      return (
-        <ul className="explore__list explore__list--projects">
-          {projects.map((p) => {
-            const authorDid = projectAuthorDid(p)
-            const meta = closure && authorDid
-              ? closure.closureByDid.get(authorDid)
-              : undefined
-            return (
-              <li key={p.uri}>
-                <ProjectListRow
-                  project={p}
-                  endorsementMeta={meta}
-                />
-              </li>
-            )
-          })}
-        </ul>
-      )
-    }
-    return (
-      <ul className="explore__grid explore__grid--projects">
-        {projects.map((p) => (
-          <li key={p.uri}>
-            <ExploreProjectCard project={p} />
-          </li>
-        ))}
-      </ul>
-    )
-  }
-
-  // certs
-  const certs = sortedCerts
-  const certDids = data.certDids
-  if (certs.length === 0) return <EmptyResults kind={kind} />
-
-  if (view === "list") {
-    return (
-      <ul className="explore__list explore__list--certs">
-        {certs.map((rec) => {
-          const did = certDids.get(rec.uri) ?? ""
-          return (
-            <li key={rec.uri}>
-              <CertListRow record={rec} did={did} showByline />
-            </li>
-          )
-        })}
-      </ul>
-    )
-  }
-
-  return (
-    <ul className="explore__grid explore__grid--certs">
-      {certs.map((rec) => {
-        const did = certDids.get(rec.uri) ?? ""
-        return (
-          <li key={rec.uri}>
-            <ActivityCard record={rec} did={did} />
-          </li>
-        )
-      })}
-    </ul>
-  )
-}
-
-function EmptyResults({ kind }: { kind: ExploreKind }) {
-  const label =
-    kind === "accounts"
-      ? "accounts"
-      : kind === "projects"
-        ? "projects"
-        : kind === "funding"
-          ? "funding receipts"
-          : "activities"
-  const icon =
-    kind === "accounts"
-      ? Users
-      : kind === "projects"
-        ? FolderGit2
-        : kind === "funding"
-          ? HandCoins
-          : CertIcon
-  return (
-    <EmptyState
-      icon={icon}
-      title={`No ${label} match`}
-      description="Try a different filter, clear the search, or pick a broader scope."
-    />
-  )
-}
-
-function sortUsers<T extends { displayName: string | null; did: string }>(
-  list: T[],
-  sort: SortOrder,
-): T[] {
-  if (sort === "alphabetical") {
-    return [...list].sort((a, b) =>
-      (a.displayName ?? a.did).localeCompare(b.displayName ?? b.did),
-    )
-  }
-  // newest/oldest don't map cleanly to actors (no createdAt on profile
-  // record here); keep insertion order which is roughly recently-indexed.
-  if (sort === "oldest") return [...list].reverse()
-  return list
-}
-
-function sortProjects<
-  T extends { value: { createdAt?: string; title?: string } },
->(list: T[], sort: SortOrder): T[] {
-  if (sort === "alphabetical") {
-    return [...list].sort((a, b) =>
-      (a.value.title ?? "").localeCompare(b.value.title ?? ""),
-    )
-  }
-  return [...list].sort((a, b) => {
-    const ac = a.value.createdAt ?? ""
-    const bc = b.value.createdAt ?? ""
-    return sort === "oldest" ? ac.localeCompare(bc) : bc.localeCompare(ac)
-  })
-}
-
-function sortCerts<
-  T extends { value: { createdAt?: string; title?: string } },
->(list: T[], sort: SortOrder): T[] {
-  if (sort === "alphabetical") {
-    return [...list].sort((a, b) =>
-      (a.value.title ?? "").localeCompare(b.value.title ?? ""),
-    )
-  }
-  return [...list].sort((a, b) => {
-    const ac = a.value.createdAt ?? ""
-    const bc = b.value.createdAt ?? ""
-    return sort === "oldest" ? ac.localeCompare(bc) : bc.localeCompare(ac)
-  })
-}
-
-/**
- * Extract the author DID from an AT-URI of the form
- * `at://<did>/<collection>/<rkey>`. Returns null on a malformed
- * URI so callers can skip the row's endorsement decoration
- * silently rather than crashing the render.
- */
-function projectAuthorDid(p: { uri: string }): string | null {
-  if (!p.uri.startsWith("at://")) return null
-  const tail = p.uri.slice("at://".length)
-  const slash = tail.indexOf("/")
-  return slash >= 0 ? tail.slice(0, slash) : null
 }
